Added notifications for Likes and Follow

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -111,12 +111,33 @@ export async function DELETE(req: Request, { params: { postId }} : {params: {pos
             return Response.json({ error: "Unauthorized" }, { status: 404 })
         }
 
-        await prisma.like.deleteMany({
-            where: {
-                userId: loggedInUser.id,
-                postId
+        const post = await prisma.post.findUnique({
+            where: {id: postId},
+            select: {
+                userId: true
             }
         })
+
+        if (!post){
+            return Response.json({ error: "Post not found" }, { status: 404 })
+        }
+
+        await prisma.$transaction([
+            prisma.like.deleteMany({
+                where: {
+                    userId: loggedInUser.id,
+                    postId
+                }
+            }),
+            prisma.notification.deleteMany({
+                where:{
+                    issuerId: loggedInUser.id,
+                    recipentId: post.userId,
+                    postId,
+                    type: "LIKE"
+                }
+            })
+        ])
 
         return new Response();
     } catch (error) {
